Done add favourites for db and show user

+++ models/registration.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, String, ForeignKey, BIGINT, Integer
 
 from db.db import Base, create_all
 
@@ -7,4 +7,4 @@ class Registration(Base):
     __tablename__ = 'registration'
     id = Column(Integer(), primary_key=True)
     user_name = Column(String(100), nullable=False, unique=True)
-    user_telegram_id = Column(Integer(), nullable=False, unique=True)
+    user_telegram_id = Column(BIGINT(), nullable=False, unique=True)
